service code cleanup
